+++ CHANGES.md
@@ -6,6 +6,7 @@ next
 
 * auth::AuthToken is now a trait, with auth::base::SimpleAuthToken being
   the first implementation.
+* ApiError::EndpointNotFound now contains the service type.
 
 0.0.4
 -----